Updated Anchor component to use the html components with the escape hatches
