+++ addons/docs/src/blocks/Anchor.tsx
@@ -11,31 +11,3 @@ export interface AnchorProps {
 export const Anchor: FC<AnchorProps> = ({ storyId, children }) => (
   <div id={anchorBlockIdFromId(storyId)}>{children}</div>
 );
-
-export function scrollToElement(element: any, block = 'start') {
-  element.scrollIntoView({
-    behavior: 'smooth',
-    block,
-    inline: 'nearest',
-  });
-}
-
-interface AnchorInPageProps {
-  href: string;
-}
-
-export const AnchorInPage: FC<AnchorInPageProps> = ({ href, children }) => (
-  <A
-    href={href}
-    onClick={(event: SyntheticEvent) => {
-      event.preventDefault();
-
-      const element = document.getElementById(href.substring(1));
-      if (element) {
-        scrollToElement(element);
-      }
-    }}
-  >
-    {children}
-  </A>
-);
